+++ Controllers/SaleOrderController.cs
@@ -1,11 +1,14 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Shop.API.Models.SaleOrderDTOs;
 using Shop.API.Services.Interfaces;
+using System.Security.Claims;
 
 namespace Shop.API.Controllers
 {
     [ApiController]
     [Route("api/saleOrder")]
+    [Authorize]
     public class SaleOrderController : ControllerBase
     {
         private readonly ISaleOrderService _saleOrderService;
@@ -18,6 +21,9 @@ namespace Shop.API.Controllers
         [HttpGet]
         public ActionResult<SaleOrderDTO> GetAllSaleOrders()
         {
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userRole != "Admin")
+                return Forbid();
             var saleOrders = _saleOrderService.GetAllSaleOrders();
             return Ok(saleOrders);
         }
@@ -25,6 +31,9 @@ namespace Shop.API.Controllers
         [HttpGet("{saleOrderId}", Name = "GetSaleOrder")]
         public ActionResult<SaleOrderDTO> GetSaleOrder(int saleOrderId)
         {
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userRole != "Admin")
+                return Forbid();
             var saleOrder = _saleOrderService.GetSaleOrder(saleOrderId);
             if (saleOrder == null)
                 return NotFound();
@@ -34,8 +43,14 @@ namespace Shop.API.Controllers
         [HttpPost]
         [Route("/create")]
         public ActionResult<SaleOrderDTO> AddSaleOrder(SaleOrderToCreateDTO saleOrderToCreateDTO)
-        {
-            var createdSaleOrder = _saleOrderService.AddSaleOrder(saleOrderToCreateDTO);
+        { 
+            var userIdClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (!int.TryParse(userIdClaim, out int id))
+                return Unauthorized();
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userRole != "Client")
+                return Forbid();
+            var createdSaleOrder = _saleOrderService.AddSaleOrder(saleOrderToCreateDTO, id);
             return CreatedAtRoute("GetSaleOrder", new { saleOrderId = createdSaleOrder.Id }, createdSaleOrder);
  
         }
@@ -43,6 +58,10 @@ namespace Shop.API.Controllers
         [HttpPut("{saleOrderId}")]
         public ActionResult<SaleOrderDTO> ChangeSaleOrderStatus(int saleOrderId, SaleOrderStatusDTO newStatus) 
         {
+           
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userRole != "Admin")
+                return Forbid();
             _saleOrderService.UpdateSaleOrderStatus(newStatus.Status, saleOrderId);
             return NoContent();
         }
@@ -50,6 +69,9 @@ namespace Shop.API.Controllers
         [HttpDelete("{saleOrderId}")]
         public ActionResult DeleteSaleOrder(int saleOrderId) 
         {
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            if (userRole != "Admin")
+                return Forbid();
             _saleOrderService.DeleteSaleOrder(saleOrderId);
             return NoContent();
         }
